Better documentation, add p2p message container type
