+++ sharding/p2p/peer.go
@@ -1,0 +1,3 @@
+package p2p
+
+type Peer struct{}
