+++ sharding/p2p/service_test.go
@@ -1,40 +1,8 @@
 package p2p
 
 import (
-	"testing"
-
 	"github.com/ethereum/go-ethereum/sharding"
 )
 
 // Verifies that Server implements the ShardP2P interface.
 var _ = sharding.ShardP2P(&Server{})
-
-func TestFeed_ReturnsSameFeed(t *testing.T) {
-	tests := []struct {
-		a    interface{}
-		b    interface{}
-		want bool
-	}{
-		// Equalality tests
-		{a: 1, b: 2, want: true},
-		{a: 'a', b: 'b', want: true},
-		{a: struct{ c int }{c: 1}, b: struct{ c int }{c: 2}, want: true},
-		{a: struct{ c string }{c: "a"}, b: struct{ c string }{c: "b"}, want: true},
-		// Inequality tests
-		{a: 1, b: '2', want: false},
-		{a: 'a', b: 1, want: false},
-		{a: struct{ c int }{c: 1}, b: struct{ c int64 }{c: 2}, want: false},
-		{a: struct{ c string }{c: "a"}, b: struct{ c float64 }{c: 3.4}, want: false},
-	}
-
-	s, _ := NewServer()
-
-	for _, tt := range tests {
-		feed1, _ := s.Feed(tt.a)
-		feed2, _ := s.Feed(tt.b)
-
-		if (feed1 == feed2) != tt.want {
-			t.Errorf("Expected %v == %v to be %t", feed1, feed2, tt.want)
-		}
-	}
-}
